Collisions being worked on
Implemented a new style of collision checking which takes into account
the corners of squares, still need to get this to work correctly with
the player/npcs. A few little neaten ups such as putting the 2 major
collisions (Block on circle and Circle on Circle) in there own functions
to make it easier to alter. Also a new feature, known as the freeze gun
is implemented which will allow the player to freeze the bots movement
(and soon shooting) for a short period of time.

Things that still need work:
#NPC Collosion with the sheild
#Collosion with the BLcoks in Y-Plane
#Being able to shoot diagonally between blocks, and inbetween them
#Bots not colliding with each othe
#The Health of each NPC as a number ontop of it
#Take you to a death screen where you can reatart after dieing
#Power up where you can sheild and shoot at the same time
#Power up where yuo sheild does melee
#NPC can clip then shoot through blocks
#Stop blocks spawning on top of each other
#Sheild morphing through Blocks
#Freeze stop bots shooting