Collisions working and other small things
The new Collision is implemented and is now working near perfect for the
player excepet for movement around corners and sliding along a surface
made of multiple blocks. The collosion detection allows for this but i
need to fine tune the reactions because currently they dont react
correctly. The collisions with bullets also needs some alterations,
because they arent fully contacting with the block before deflecting and
are also not performing correctly on corners.

Small aditives include that now NPCs display there health as a number
above them, You can now change the game to show solid objects and
bullets having a delete time (because of lag when too many bullet
objects exist). Now the fereze gun also stops the NPCs shooting as well
as moving but has been shortened to compensate (now 10seconds from 20
seconds). Some minor re-arrangement of ordering to make it look slightly
neater (But Didnt really do anything)

Things to be Completed:
#NPC Collosion with the sheild
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#Power up where you can sheild and shoot at the same time
#Power up where yuo sheild does melee
#Stop blocks spawning on top of each other
#Freeze stop bots shooting
#Add a shift to dash function
#Fix that if you click players exact center crash due to dividing by
zero
#NPCs can still morph through blocks, problem known need fix