Borders and NPC Non-Working Block dodging!!!
Been doing some minor experimenting with apperances and like the look of
basic filled shapes with black borders, though with 2 pixel borders
apparenntly pygame cant handle the task misses some pixels in the
border. Besided this some toying with colours and other small changes.
As far as mechanics go there has been a slightly better system for the
NPC collision with the sheild added, but currently not working correctly
and they bouce off it insterad of being moved to the edge (Though im
thinking i could stick with this as it will make it easier to move
through crowds of enemies without as large risk of your sheild getting
quickly destroyed.  The SMG also got moved to the new shoot delay system
with the rest of the players weapons which should make it feel more
reactive on the first bullet. The basics of NPC not getting stuck on
blocks has been added but currently does nothing so it is useless for
now but hopefully can be made to work.

Needs work:
#NPC Collosion with the sheild
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#Maybe add a pause menu
#Power up where you can sheild and shoot at the same time
#Power up where yuo sheild does melee
#Stop blocks spawning on top of each other
#Add a shift to dash function
#Fix that if you click players exact center crash due to dividing by
zero
#NPCs can still morph through blocks, problem known need fix
#Find a prettier way to call the npcs (currently looks pretty munted)
#Get better colours
#try adding some basic NPC Ai for transversing blocks, eg if stuck
change y velocity
#add a new way of spawning enemies in a wave style system
#shotgun triginometery (and maybe add 5 bullets instead of 3)