Squashing Blocks and Dashing
Besides a few minor bug fixes the main changes are that now when the
player places blocks ontop of npc, the npc will now die. I thought that
this could add some interesting aspects of gameplay as now the player
has to carefully choose how they use there blocks. The other thing thats
added is that now there is a dash function for the player (As per
request if Jack) which is initiated when the left shift key is pressed,
and is king of fun as the player rebounds of blocks like the bullets.
Planning to add friction into this function in the future.

Needs Adding:
#NPC Collosion with the sheild
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#Power up where you can sheild and shoot at the same time
#Power up where yuo sheild does melee
#POwer up where you can move on blocks
#Stop blocks spawning on top of each other
#NPCs can still morph through blocks, problem known need fix - The bot
movement is after the collision
#Find a prettier way to call the npcs (currently looks pretty munted)
#try adding some basic NPC Ai for transversing blocks, eg if stuck
change y velocity
#add a new way of spawning enemies in a wave style system
#shotgun triginometery (and maybe add 5 bullets instead of 3)
#friction for the dash function