Power-ups and NPC block dodging
Now there is currently one Powerup in the game with allows the player to
shoot and operate the sheild at the same time for a limited time. This
works by dropping an obeject and then the player has to go and pick it
up in order for it to initiate. As well as this there is the start of
basic block dodging for the npcs, hopefully so that they wont get stuck
on so many obsticals.

needs adding:
#NPC Collosion with the sheild
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#Power up where yuo sheild does melee
#POwer up where you can move on blocks
#NPCs can still morph through blocks, problem known need fix - The bot
movement is after the collision
#try adding some basic NPC Ai for transversing blocks, eg if stuck
change y velocity
#add a new way of spawning enemies in a wave style system
#shotgun triginometery
#friction for the dash function
#Multiplayer
#Create big map with following camera (and spawn structers and stuff)
#powerups lasting longer than they should
#Adding a timer for the freezeray cooldown
#Balancing the wepaons