Sheild Melee Power-up and freeze-timer
A few minor additives including a second power up which allows the
player to use their sheild as a melee weapon, the sheild changes colours
as a visual reperesntaion of whether or not the power-up is active (As
if it isnt the bots will be harming the players sheild!). Also the
Freeze gun now has a timer accompanying it so the player can till if
they can use it or not, it counts down from 30 to zero, if it is zero
the player can us it and it will reset. Another little thing added to
the freeze-ray is that after being fired it will automaticaly return
itself to the wepaon previsously being used, just to make it a little
easier, Note that this does not applie if the can is in cool down and
the player will be stuck with a useless weapon. Besides this there where
a few minor bug fixes with the firing mechinism with the first powerup
fixed and other minor adjustments. This comes close to concluding all
the core game mechainics being added to the game, as if anymore are
added the game will become much to cluttered, (as it already is
beggining to with it being near impossible to effecitivly balance all
the different mechanics as is!!!)

Needs adding/ fixing:
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#POwer up where you can move on blocks
#NPCs can still morph through blocks, problem known need fix - The bot
movement is after the collision todo witrh the movement occuring after
the block thing occurs
#add a new way of spawning enemies in a wave style system
#shotgun triginometery
#friction for the dash function
#Multiplayer
#Create big map with following camera (and spawn structers and stuff)
#Balancing the wepaons
#powerup 2 doesn't work during freezing
#stop bots spawning where the player does
#Sheild collision with the blocks (currently most bugged thing in the
game, and thats saying something!!!)