Battle moon, enough said.
Now there is a third power up that can be received which is a small moon
which orbits the player and shoots bots. Unlike the other 2 power ups
this one doesnt have a timer but rather health, so it is the players
responability to keep it alive. I am thinking of making it so that is
multiple powerup 3 tokens a received it increases the level of the bot,
eg level 1: smg, low health level 2: shot-gun, mid health ect till the
max level where it can shoot multiple bullets at different bots. Besides
this there have just been some minor clean up in the code to make it
look a little more pretty, but it still remains rather ugly.

Needs adding:
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#shotgun triginometery
#Multiplayer
#powerup 2 doesn't work during freezing
#Player Collosion with enemies
#Block blocking turns out to be broken from the upgrade to the new
collisions, need to reinstall
#sliding across multiple blocks doesnt work (because the player
asctually moves inside the block)
#Collisions with the corners of blocks (Should be taken into account
with the new Collisions but isnt)
#wave system kinda in but not happy with it
#power up that puts a ball that spins around the player
#make the battle moon so that it shoots at the closest npc
#make the battle moon have different levels (with different
colours/weapons/health)