Bit of a clean up and better player freindly AI
Found a few unesseceray sections of code which where just wasting space
and so cleaned them up so that now all the main object classes have been
merged into 1 class (eg. blocks, turrets and powerups). Also i decided
that the way the Player freindly AI was stupid as it didnt garantee that
it would shoot at the closest enemy but rather the one highest up on the
entity list. This has been fixed so that now turrets and the battlemoon
shoots at the enemy cloesest to itself. Also a bug which meant that
occasionaly the powerups would spawn inside of blocks has been fixed,
now all of blocks, turrets and powerups work on the same system of
veiwing the screen as a grid and checks if there is a available space or
not, Note that this might be quicker to do if a map matrix was used of
1s and 0s (Close to Jacks system of collisions in his particle
simulator). Besides this no major alterations have occured.

Needs adding/Fixing:
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#shotgun triginometery
#Multiplayer
#powerup 2 doesn't work during freezing
#Player Collosion with enemies
#sliding across multiple blocks doesnt work (because the player
asctually moves inside the block)
#Collisions with the corners of blocks (Should be taken into account
with the new Collisions but isnt)
#wave system kinda in but not happy with it
#Add different levels of Turrets
#enemy with homing bullets
#a function for picking available x and y positions
#make it so you have to win each weapon
#placing system works through a matric of coordinates instead of for
loops.