Acquiring Currency and Shopping
I decided that the current system of being given stuff imeditatly was a
bit to easy and took away the excitment (If there is a possible way to
feel any emotion about Game.py) of deiscovering new things. So i have
decided to add a shop system.  It is as basic as it gets, by killing a
bot you get $10 (I know dont spend it all at once). This can then be
taken to a store (A blue block) and cashed in for new weapons, health
and turrets.

Need Fixing/Adds:
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#shotgun triginometery
#Multiplayer
#powerup 2 doesn't work during freezing
#sliding across multiple blocks doesnt work (because the player
asctually moves inside the block)
#Collisions with the corners of blocks (Should be taken into account
with the new Collisions but isnt)
#wave system kinda in but not happy with it
#Add different levels of Turrets
#enemy with homing bullets
#a function for picking available x and y positions
#all circle on circle collisions
#Finish the shop and add new items
#Draw the weapons
#Make the money move towards the player (Just looks cool) Note probs use
a 1/x^2 graph (eg. moves faster the closer it gets)