Money Magnet, Neater Shop (Barley)
I added in the movement of money towards the character, and i think it
looks quite cool, I was tossing up whether to make the movement in the
style of a 1/x graph (Though I experimented with a 1/sqrt(x) but found
the change in velocity is it grew closer was to minor) besides real
magnetism bein cloder to a 1/x^2 graph I beilve just because it appeared
the most natrual form of movement. I also did some quick touch ups on
the shop, now the prices are closer to what they really will be and the
prices appear on the clicking block.
Also the start of a Homing bullet weapon is writen but im thinking of a
way of adding them without adding a new class (I guess just through
adding a new Function in the existing way is the most logical) but I'll
need to have a proper look at it.

Needs Adding/Fixing:
#Bots not colliding with each other
#Take you to a death screen where you can reatart after dieing
#shotgun triginometery
#Multiplayer
#powerup 2 doesn't work during freezing
#sliding across multiple blocks doesnt work (because the player
asctually moves inside the block)
#Collisions with the corners of blocks (Should be taken into account
with the new Collisions but isnt)
#wave system kinda in but not happy with it
#Add different levels of Turrets
#enemy with homing bullets (Or a player weapon)
#all circle on circle collisions
#Finish the shop and add new items (Maybe add Ammo)
#Draw the weapons
#A way to tell what weapon is selected