Did you think I was doing nothing?
Well though there hasnt been any updates for the last bit its not been
for the lack of work, during the last week (or so) I have completely
reworked the whole game from scratch as it was becoming difficult for me
to follow or even find what i wanted. The new code does look a little
better and gone are the old dictionaries and replaced with matrices so
that now i can deal with levels of things much easier than having an 8
set of elifs. The code is fully working but there is a few new things
added, such as an actual wave spawn system and a few othere little
things. The shop hasnt been re-added but im a little reluctant as last
time as it ended up as crazy hard-coding (enough to put diamond to
shame) but besides that an a couple of bugs the old game.py is on its
way back