Still not back to its former glory
Getting closer to being back to a proper game, but i still have to
re-add the user interface and the shop both of which i have been putting
of because of the previous swamp of spaghetti that the shop section
turned off. But I have added in a new scroll wheel system to change
weapons for convieence, but i am thinking of doing it so it only changes
between the primary weapons because finding a single weapon it a set of
6 can be a little frustrating. I have also added a title add the start
of each wave to have it easier for the player to track the number of
waves they have survived and what bots they should expect on the next
wave. There is still a bug with the shooting of the bots that I am
having touble tracking (possibly my autism) but hopefuly after I find it
the is the majority of the movement mechanics done.