In working order (well kinda)
found that bug (and it was my autism) and now the game is working the
majortiy of the time except for the excpetion where it is getting a
error "Python int too large to convert to C long" which is going to be
my next thing on my list of things to fix. I also added the players
health and fixed a few minor issues realted to the powerup/sheild system
which meant the player could always sheild'n'shoot. Besides this no
major additives/changes