Shakey screen and homing bullets
Well as the title says there has been a couple cool things added. Now
the screen shakes whenever the player is it by a bullet or a melee
atack. There is also homing bullets available as a weapon for the
player, but there is a small bug where as the bullets approach vertical
movement the bullets get stuck in a small oscilaating type movement.
Besides this a few small bug fixes and now powerup pickup objects share
the same colour as the sheild colour they produce.