Sound Effects
I have added in some basic sound effects for weapons and the like, it
adds a certain charm to the game.
I also fixed a bug due to weapon changing which didnt allow for weapons
to be changed with the keys. Now both key changing and mouse wheel
changing of weapons work