Nothing Big
I have just done a little more cleaning up of the code and turned a
couple more of the image sections into dictionaries to make it a little
neater. Also after much berating and bullying I have increased the
frame-rate from 30fps to 60fps and in the process have also slightly
increased the speeds of everything in an attempt to make the game feel
more face paced, hoping that this will make the game fell more lively.